moved softmax function and updated prev in Tensor
