+++ CMakeLists.txt
@@ -8,7 +8,8 @@ include_directories(src)
 
 set(SOURCES
     main.cpp
-    src/gradients.cpp
+    src/backwards.cpp
+    src/functions.cpp
 )
 
 set(HEADERS
@@ -36,16 +37,20 @@ if(BUILD_TESTING)
     add_executable(array_test tests/array_test.cpp)
     add_executable(linear_test tests/linear_test.cpp)
     add_executable(gradient_test tests/gradient_test.cpp)
+    add_executable(backwards_test tests/backwards_test.cpp)
+    add_executable(functions_test tests/functions_test.cpp) # Add the new functions test
 
     target_link_libraries(tensor_test gtest gtest_main)
     target_link_libraries(array_test gtest gtest_main)
     target_link_libraries(linear_test gtest gtest_main)
     target_link_libraries(gradient_test gtest gtest_main)
-
+    target_link_libraries(backwards_test gtest gtest_main)
+    target_link_libraries(functions_test gtest gtest_main) # Link the new test with gtest
 
     add_test(NAME tensor_test COMMAND tensor_test)
     add_test(NAME array_test COMMAND array_test)
     add_test(NAME linear_test COMMAND linear_test)
     add_test(NAME gradient_test COMMAND gradient_test)
-
+    add_test(NAME backwards_test COMMAND backwards_test)
+    add_test(NAME functions_test COMMAND functions_test) # Add the new test to CTest
 endif()
